feat(adoption): scope Java proof to nested workspaces
