+++ tests/fixtures/adoption_repos/mixed_nested_java_workspaces/services/catalog/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    java
+}
