+++ tests/fixtures/adoption_repos/mixed_nested_java_workspaces/services/catalog/gradlew
@@ -1,0 +1,2 @@
+#!/usr/bin/env sh
+exec gradle "$@"
