+++ tests/fixtures/adoption_repos/mixed_nested_java_workspaces/services/orders/mvnw
@@ -1,0 +1,2 @@
+#!/usr/bin/env sh
+exec mvn "$@"
