+++ tests/fixtures/adoption_repos/mixed_nested_java_workspaces/services/orders/pom.xml
@@ -1,0 +1,6 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>example.com</groupId>
+  <artifactId>orders</artifactId>
+  <version>1.0.0</version>
+</project>
